pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = "RAG Agent with Human-in-the-Loop for Customer Support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.0.1,<5.0.0",
    "bs4>=0.0.2",
    "chromadb>=1.4.1",
    "fastapi[standard]>=0.128.0",
    "langchain>=1.2.7",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.7",
    "langgraph-checkpoint-postgres>=3.0.3",
    "langgraph-cli[inmem]>=0.4.12",
    "passlib[bcrypt]>=1.7.4",
    "psycopg-binary>=3.3.2",
    "psycopg-pool>=3.3.0",
    "pydantic-settings>=2.12.0",
    "redis>=6.4.0",
    "sqlmodel>=0.0.32",
]
